[global-30] add profile info
