+++ src/Navigation.tsx
@@ -1,13 +1,17 @@
 import AppBar from "@material-ui/core/AppBar";
+import Avatar from "@material-ui/core/Avatar";
 import Box from "@material-ui/core/Box";
+import Button from "@material-ui/core/Button";
 import Link from "@material-ui/core/Link";
+import Menu from "@material-ui/core/Menu";
+import MenuItem from "@material-ui/core/MenuItem";
 import { makeStyles } from "@material-ui/core/styles";
 import Toolbar from "@material-ui/core/Toolbar";
 import Typography from "@material-ui/core/Typography";
 import * as React from "react";
+import { useState } from "react";
 import { Link as RouterLink } from "react-router-dom";
-import { useStoreState } from "./config/store";
-import User from "./User";
+import { useStoreActions, useStoreState } from "./config/store";
 
 const useStyles = makeStyles(theme => ({
   root: {
@@ -29,10 +33,21 @@ const useStyles = makeStyles(theme => ({
   }
 }));
 
-const Navigation = () => {
+function Navigation() {
   const year = useStoreState(state =>
     state.global.preferences.year !== null ? state.global.preferences.year : ""
   );
+  const userLogout = useStoreActions(actions => actions.global.userLogout);
+  const user = useStoreState(state => state.global.user);
+  const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null);
+
+  const handleClick = (event: React.MouseEvent<HTMLButtonElement>) => {
+    setAnchorEl(event.currentTarget);
+  };
+
+  const handleClose = () => {
+    setAnchorEl(null);
+  };
   const classes = useStyles();
 
   return (
@@ -60,11 +75,46 @@ const Navigation = () => {
               {year}
             </Typography>
           </Box>
-          <User />
+
+          <Button
+            aria-controls="user-menu"
+            aria-haspopup="true"
+            onClick={handleClick}
+          >
+            <Avatar
+              alt={(user !== null && user.displayName) || ""}
+              src={(user !== null && user.photoURL) || ""}
+            />
+          </Button>
+          <Menu
+            id="user-menu"
+            anchorEl={anchorEl}
+            keepMounted
+            open={Boolean(anchorEl)}
+            onClose={handleClose}
+            getContentAnchorEl={null}
+            anchorOrigin={{
+              vertical: "bottom",
+              horizontal: "center"
+            }}
+            transformOrigin={{
+              vertical: "top",
+              horizontal: "center"
+            }}
+          >
+            <MenuItem
+              component={RouterLink}
+              to="/profile"
+              onClick={() => handleClose()}
+            >
+              Profile
+            </MenuItem>
+            <MenuItem onClick={() => userLogout()}>Logout</MenuItem>
+          </Menu>
         </Toolbar>
       </AppBar>
     </div>
   );
-};
+}
 
 export default Navigation;
